Pipeline Update
This new update introduces the Pipeline! This offers a way to interact with entire collections of Result functional statements in a clear and composable manner. Allowing you to establish Timeouts and Retry failed results. Essentially creating a composable method that is failure-aware.
The Linq extension methods for it allows you to use Linq expressions against it providing a clean and readable method for combining the results of Pipelines which has been something I've been wracking my brain for how I can cleanly introduce such a feature. Even if the extension methods on their own can be a bit confusing to a new developer but I feel that sacrifice of learnability is worth it.
I'm quite pleased with how it turned out even if it deserves a few more extension methods. I'm leaning more towards the execution of the Result pattern rather than sticking to the pattern as a whole as I feel like that is where it shines most. Its the supporting actor rather than the star.